% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglabeller.R
\name{gglabeller}
\alias{gglabeller}
\title{gglabeller}
\usage{
gglabeller(gg, mapping = gg$mapping, data = gg$data, geom = c("text",
  "label"), ..., width = "100\%", height = "100\%")
}
\arguments{
\item{gg}{ggplot object}

\item{geom}{geom to use -- "text" or "label"}

\item{...}{additional arguments to pass to \code{\link[ggrepel]{geom_text_repel}} or
\code{\link[ggrepel]{geom_label_repel}}}

\item{width}{width of plot}

\item{height}{height of plot}
}
\description{
Shiny gadget for selecting points to label in a ggplot object
}
\details{
Uses ggrepel for controlling position of labels. Parameters tab of
widget can be used to adjust ggrepel parameters.
}
\examples{

if (interactive()){

  p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()

  gglabeller(p, mapping = aes(label = rownames(mtcars)))

}
}
\author{
Alicia Schep
}
