% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglabeller.R
\docType{data}
\name{gglabeller_example}
\alias{gglabeller_example}
\title{gglabeller_example}
\format{list with three elements: the plot, the indices for rows of data
that are labelled, and the code to regenerate the plot}
\usage{
gglabeller_example
}
\description{
Example output from gglabeller function
}
\keyword{datasets}
