% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglabeller.R
\name{gglabeller}
\alias{gglabeller}
\title{gglabeller}
\usage{
gglabeller(gg, mapping = gg$mapping, data = gg$data, geom = c("text",
  "label"), ..., width = "100\%", height = "100\%")
}
\arguments{
\item{gg}{ggplot object}

\item{mapping}{aesthetics mapping, from \code{\link[ggplot2]{aes}}}

\item{data}{data.frame with data for plotting}

\item{geom}{geom to use -- "text" or "label"}

\item{...}{additional arguments to pass to \code{\link[ggrepel]{geom_text_repel}} or
\code{\link[ggrepel]{geom_label_repel}}}

\item{width}{width of plot, as percent or pixels}

\item{height}{height of plot, as percent or pixels}
}
\value{
Upon clicking the "Done" button, the gadget returns a list with three elements:
\describe{
\item{plot}{The plot itself}
\item{ix}{The indices of the rows of the data used for labelling}
\item{code}{a character vector with a code snippet for recreating the plot reproducibly}
}
}
\description{
Shiny gadget for selecting points to label in a ggplot object
}
\details{
Click on a point or brush over a set of points to label them. Clicking
on a labelled point or brushing over a set of points that are all labelled
will remove the labels. Brushing over some points that are labelled and some
that are not will add labels to the ones that are not.

gglabeller sses ggrepel for controlling the position of labels. The parameters tab
can be used to adjust ggrepel parameters.
}
\examples{

if (interactive()){

  p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()

  plabelled <- gglabeller(p, mapping = aes(label = rownames(mtcars)))

}
}
\author{
Alicia Schep
}
